% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_course}
\alias{default_course}
\title{Default coreR course}
\format{
## `default_course`
A data frame with 21 rows and 2 columns:
\describe{
  \item{module}{Module name}
  \item{course}{The name of the course file}
}
}
\usage{
default_course
}
\description{
A set of lessons ordered according to coreR October 2024 course.  See also
`example_course`
}
\keyword{datasets}
