% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_course}
\alias{example_course}
\title{Example coreR course}
\format{
## `example_course`
A data frame with 10 rows and 2 columns:
\describe{
  \item{module}{Module name}
  \item{course}{The name of the course file}
}
}
\usage{
example_course
}
\description{
A subset of lessons ordered according to coreR October 2024 course.  See also
`default_course`
}
\keyword{datasets}
